Refactor GREL Get tests
- move helper up to RefineTest
- move tests to the correct module
